Moved classrooom (3 o) module to classroom
